add macro for check with exception (#40)
Summary:

With previous approach using `CUDACHECKGOTO`|`NCCLCHECKGOTO`, it required `ncclResult_t res` variable and a goto label. C++ compiler complains `res` is set but not used.

It is more straightforward to just throw the exception within the CHECK macro.

This patch defines `CUDACHECKTHROW` and `NCCLCHECKTHROW` macros for error checking in ctran constructors.

Reviewed By: cenzhaometa

Differential Revision: D53187428