NCCL logger to enable asynchronous logging
Summary:
Current oilfs used for GenAI/LLama3 uses blocking IO that causes inflated latency during NCCL message logging, because the logging is conducted in the critical path of NCCL runtime. We found that this delay in logging causes the perf regression by ~23%, measured by the end-to-end job WPS (xlf.train.wps) as shown in below figures:

https://pxl.cl/40DQP: with NCCL logging
https://pxl.cl/40nhf: without NCCL logging
Thus, we have to either be pending on the improvement on oilfs or switch to other file systems like locate fs. Or we have to turn off NCCL logging completely, which in turn limits our troubleshooting capability.

This diff is a quick remedy to add support for asynchronous logging of NCCL messages, by removing logging from the critical path altogether. It should be compatible with oilfs or any existing file systems.

Reviewed By: YulunW

Differential Revision: D52264046

fbshipit-source-id: 14695bbf33b0cae3dae879e11d2bf911c2c98ae5