Fix issue of logging in destructor
Summary:
Previously in async mode, logging would invoke NcclLogger::getInstance() to get the singleton of NcclLogger and then call the log() function of the singleton. This means we need to ensure any logging happened before the destructor of the NcclLogger singleton being called. Otherwise, we will get a segfault.

Usually this is not a problem, but it would be if we are logging in the destructor of other static objects. Like by the meta-tuner D53686248.

The diff does the following to fix this issue:
1. make the singleton a class variable.
2. make log() static so that logging doesn't require getting an instance of NcclLogger in `debug.cc`. log() will check whether the singleton is still valid, and log to file directly if the singleton is not valid.
3. make log() noexcept so it is safe to call in the destructor.
4. initialize the singleton in ncclDebugInit() rather than in the first call to log().

The change is made in both 2.18 and 2.19

Reviewed By: minsii

Differential Revision: D54411382

fbshipit-source-id: 3d95873a86ce264f33d0b6cd64f9c03fdebbcb57