allow socketTryAccept to ignore EINTR error
Summary:
In Ctran, we have a separate thread waiting for incoming RoCE connection request, via a blocking wait in socket accept. When running with some models, we found other module may trigger signal and caused EINTR returned from socket accept. EINTR was treaded as error before and caused the waiting thread fail.

This patch allows EINTR to be ignored so that the thread can continue waiting in accept.

Reviewed By: wesbland

Differential Revision: D54465580

fbshipit-source-id: 674fdb42d6a157685fb4e4f35beb0ab5b3c3d116