Bug-fix in graph/paths.cc CVAR
Summary:
When converting to CVARs, we introduced a bug in D52454684 based on
type conversion from int64_t to int.  The code assumed that casting
from int64_t to int would simply discard the high-order bits, but this
is implementation defined and should not be assumed.  This patch
directly uses "int" types for the CVAR to avoid such issues.

Reviewed By: kingchc

Differential Revision: D53170187

fbshipit-source-id: 495a1b5a034fc3e24e7f06f9e55a1fa8cf809c3e