ncclResult_t return error a string for better understanding of the error from the logs
Summary: This diff is part of the NCCL-EXP project and aims to improve the observability of errors in the logs. The changes made in the code include adding a call to `ncclGetErrorString` to convert the `ncclResult_t` return value to a string for better understanding of the error. The changes were made in the `checks.h` file and the `WARN` function was modified to include the error string. The changes were made to the `nccl-exp` project and version `v2.18.3-1`.

Reviewed By: minsii

Differential Revision: D54634218

fbshipit-source-id: dfaa69a85c9935eea25520015d2e6a6111659240